refactor(git): generic refactor
Started thinking about a new problem relating to how to deal with a
series of operations, as quick has been doing so far.

Now that we are trying to react to errors, it would be cool if we could
stop early, but this require maintaining a list of things that have
failed to avoid running them, and so we need to perform that logic
inside of the function (at least it's the more reasonable approach imo)

One way I can sketch rn might be to copy the hashmap structure of repos
as a reference, and then delete the references from our cloned map as
they fail. This is not ideal in any specific way, but it's a good first
approximation.

Signed-off-by: Christina Sørensen <christina@cafkafk.com>