feat(git): started adding multi instruction logic
This follows up on previous commits message, but with a completely
different implementation that is a lot tighter.

It doesn't handle commit messages, but quick is now implemented.

This has also shown that the previous fall-through behaviour of quick
wasn't a bug but a feature, as it would try to pull, not fail if it
couldn't.

This means it would be a good idea to implement an alternative version
of quick that keeps this old behaviour, or probably more likely,
creating a new command that does what the quick in this commit does, and
then finding a way to recreate the old behaviour of quick.

This is dev branch thou, so I don't have to do that before pushing :D

Signed-off-by: Christina Sørensen <christina@cafkafk.com>