doc: added roadmap
Added a tentative roadmap for release 0.1.0 as well as a future planned
roadmap.

This is mostly to help myself not trying to add too many features in
between releases as dev is already far ahead of main.

Ideally, I'd like to iterate more rapidly on main. Given the basic
structure of the config is slowly stabilizing, this will be less painful

Signed-off-by: Christina Sørensen <christina@cafkafk.com>