fix!: fixed quick, fast messages
Fixed quick fast messages by "leaking" the memory for the input
effectively makiing it static.

This feel extremely hacky, and should never be used in a loop.

Consider this temporary until I find a better way, and or smart enough
people to tell me it's not as bad as I feel it is.

...Sorry Djikstra.

Signed-off-by: Christina Sørensen <christina@cafkafk.com>