fix: fixed commit with editor regression
This allows committing with the users set editor, which is very handy if
you don't want to actually make meaningful commits.

This existed before, but implementing the instruction architecture for
composite actions broke it, because the status code refactor used
output() instead of status().

Thankfully, this does support exit codes still, but it does create some
cluttered output which is not terrible great UX, but this is not
extremely important, as this associated function isn't even currently
used by any user facing commands (except its explicit subcommand).

:D

Signed-off-by: Christina Sørensen <christina@cafkafk.com>