feat(benches): added benchmarking
Setup only for loading config file for now, probably not a lot of code
here that is super intensive, but having the structure set up to bench
is pretty nice when it's needed suddenly.

Two steps ahead as always, still haven't capitalized on the architecture
rewrite, yet here I am adding this.

Signed-off-by: Christina Sørensen <christina@cafkafk.com>