refactor: started work on generic macro
This is to help make series operations more generic.

This refactor has been needed for a while, and I'm very excited to
finally start working on it, but this is just the bare minimum, we
still need to handle on a repo type basis.

But also, yes, this starts the slow crawl towards generic repos intead
of hardcoded functions for each repo type (links, gitrepo) etc

Signed-off-by: Christina Sørensen <christina@cafkafk.com>